pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wshawk"
version = "4.0.0"
authors = [
    {name = "Regaan"}
]
description = "Professional WebSocket security scanner with real vulnerability verification, session hijacking tests, and CVSS scoring"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
keywords = ["websocket", "security", "scanner", "penetration-testing", "bug-bounty", "vulnerability", "xss", "sqli", "session-hijacking", "cvss", "playwright", "oast", "waf-bypass"]
dependencies = [
    "websockets>=12.0",
    "playwright>=1.40.0",
    "aiohttp>=3.9.0",
    "PyYAML>=6.0",
    "flask>=3.0.0",
    "uvicorn>=0.25.0",
    "fastapi>=0.109.0",
    "python-socketio>=5.10.0",
    "dnspython>=2.4.0",
    "python-whois>=0.8.0",
    "cryptography>=41.0.0",
]

[project.urls]
Homepage = "https://github.com/regaan/wshawk"
"Bug Reports" = "https://github.com/regaan/wshawk/issues"
Source = "https://github.com/regaan/wshawk"

[project.scripts]
wshawk = "wshawk.__main__:cli"
wshawk-interactive = "wshawk.interactive:cli"
wshawk-advanced = "wshawk.advanced_cli:cli"
wshawk-defensive = "wshawk.defensive_cli:cli"

[tool.setuptools]
packages = {find = {}}

[tool.setuptools.package-data]
"wshawk" = [
    "payloads/*.txt",
    "payloads/**/*.json",
    "web/templates/*.html",
    "web/static/*"
]
